Handle python 3's change to default close_fds=True
Change mux to explicitly specify close_fds=False to deal with the fact
that python apparently changed the default in 3.2, and it absolutely
does need to keep the fds open.

Add close_fds to the other subprocess-related functions: readpipe,
batchpipe, and exo, but default to True, since it doesn't look like we
need to preserve the open fds in our other calls.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>