Only change LC_CTYPE for bup-python itself, not sh or subprocesses
With the previous approach, bash would complain on systems that don't
provide the ISO-8859-1 locale (python always supports it), and those
complaints would break some tests.

In any case, since the original hack was only intended to affect
bup-python, try to make that the case.  Specify the LC_CTYPE only to
python via env (so bash won't see it), and then undo the environment
changes at the top of cmd/bup so they won't affect unrelated
subprocesses.

Thanks to Johannes Berg for reporting the problem and helping devise
the fix.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>